country combo ui fixed

+++ WereViewApp/Models/POCO/IdentityCustomization/CountryDetectByIP.cs
@@ -4,8 +4,8 @@ namespace WereViewApp.Models.POCO.IdentityCustomization {
     public class CountryDetectByIP {
         [Key]
         public int CountryDetectByIPID { get; set; }
-        public int BeginingIP { get; set; }
-        public int EndingIP { get; set; }
+        public long BeginingIP { get; set; }
+        public long EndingIP { get; set; }
 
         public int CountryID { get; set; }
